started new circular buffer addressing scheme
We are computing all of the element delays relative to the phase center
of the array; To handle this in the circular buffers, the phase center
reference needs to lag behind the write pointer by half of the maximum
delay across the array. The read pointer can then be offset from phase
center by the element delay that is computed relative to the phase
center.