change interpolation/decimation filters to CIC filters
FIR filters were being used previously, and it turns out the filter
lengths were way too long, requiring way too many multiplications.
This caused all 32 GB of RAM plus 16 GB of swap on my machine to
be used when compiling a delay and sum beamformimg Quartus project...

CIC filters don't require any multiplications, which is much more efficient.
FIR filters are used to compensate for the frequency response of the CIC
filters, and shifts are used to compensate for the gain of the CIC filters.