move delay computation out of the dataplane
the idea here is that we'd rather do trig calculations in the software.
the register interface has changed to delays instead of angles, but the
software interface will still be angles. MATLAB's C Coder can be used
to generate code to convert angles to delays; this won't integrate with
our current device driver scheme, so some changes are imminent there.