fix implementation of mean and variance calculations
The initial conditions were not correct for the exponential moving
average and variance. The exponential moving variance was not
implementing the difference equation properly at all...

The output matches the MATLAB code much better now. The exponential
moving variance seems to overestimate the variance during periods of
small variance; this is causing excess noise in "silent" periods
compared to the MATLAB code.