Refactored Autogen Framework. Updated echo and bitcrusher models
The Autogen framework now lives in the autogen folder and its interfaces
are described by the README in the same folder. It is mostly the same as
before but now with new configuration and no need for the framework to
be pasted into each model directory.

The configuration is divided into
`model.json` which is used by the Autogen tools and the framework and
`modelparameters.m` which is extra configuration just pertaining to the
framework running in Matlab/Simulink.

Additionally, the refactoring to `sm_stop_process_output` have resulted
in significant performance gains. As well, zero data is no longer
generated between audio samples.

The echo and bitcrusher models have been updated to use the new
framework.