Initial support for vectorized models without the Avalon interface
The mp.useAvalonInterface flag was added to maintain compatibility with
the old models that still use the Avalon interface is simulation.

AudioSource objects can now be returned as time series objects, which
allows them to be used in "From Workspace" blocks in Simulink.

Echo model is now vectorized; initial testing suggests that the delay is
still working as before. DPRAM does not accept vector inputs, so
mux/demux was used.